Move repository interface
